+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 All notable changes to this project will be documented in this file.
 
 ## [0.2.0] - 2025-01-??
+###
+- Added bus scan function to examples/cli.rs
+
 ### Changed
 - Moved away from opencbm bindings to A Rust native xum1541 implementation (xum1541)
 - Added Cbm::load_file_ascii() and Cbm::load_file_petscii()